test(cli): consolidate Ollama inventory coverage
