+++ src/lib/inference/local.test.ts
@@ -24,6 +24,16 @@ function emptyOllamaInventoryCapture(): () => string {
   };
 }
 
+function makeOllamaCapture(responses: ReadonlyArray<{ match: RegExp; output: string }>) {
+  const calls: (readonly string[])[] = [];
+  const capture = ((command: string | readonly string[]) => {
+    const argv = typeof command === "string" ? [command] : command;
+    calls.push(argv);
+    return responses.find(({ match }) => match.test(argv.join(" ")))?.output ?? "";
+  }) as Parameters<typeof getOllamaModelOptions>[0];
+  return { capture, calls };
+}
+
 import {
   buildOllamaProbeOptions,
   CONTAINER_REACHABILITY_IMAGE,
@@ -45,6 +55,8 @@ import {
   isLocalProviderProbeOutputHealthy,
   isOllamaRunnerCrash,
   LOCAL_INFERENCE_SANDBOX_HOST_URL_ENV,
+  OLLAMA_HOST_DOCKER_INTERNAL,
+  OLLAMA_LOCALHOST,
   parseOllamaList,
   probeLocalProviderHealth,
   probeOllamaAuthProxyHealth,
@@ -931,44 +943,66 @@ describe("local inference helpers", () => {
     expect(parseOllamaList("NAME ID SIZE MODIFIED\n\n")).toEqual([]);
   });
 
-  it("returns parsed ollama model options when available", () => {
-    let call = 0;
-    const mockCapture = () => {
-      call += 1;
-      return call === 1
-        ? JSON.stringify({ models: [] })
-        : "nemotron-3-nano:30b  abc  24 GB  now\nqwen3:32b  def  20 GB  now";
-    };
-    expect(getOllamaModelOptions(mockCapture)).toEqual(["nemotron-3-nano:30b", "qwen3:32b"]);
+  it("returns no models when the Windows host reports an empty inventory", () => {
+    setResolvedOllamaHost(OLLAMA_HOST_DOCKER_INTERNAL);
+    const { capture, calls } = makeOllamaCapture([
+      { match: /\/api\/tags/, output: JSON.stringify({ models: [] }) },
+    ]);
+    expect(getOllamaModelOptions(capture)).toEqual([]);
+    expect(calls).toHaveLength(1);
+    expect(calls[0].join(" ")).toContain(`http://${OLLAMA_HOST_DOCKER_INTERNAL}:11434/api/tags`);
   });
 
-  it("does not fall back to the CLI after a malformed Ollama inventory", () => {
-    const mockCapture = vi.fn(() => JSON.stringify({ models: [{}, { name: "qwen3.5:9b" }] }));
-    expect(() => getOllamaModelOptions(mockCapture, () => {})).toThrow(
-      /Could not read Ollama models from 127\.0\.0\.1:11434 after 3 attempts/,
-    );
-    expect(mockCapture).toHaveBeenCalledTimes(3);
+  it("falls back to `ollama list` on loopback when /api/tags is empty", () => {
+    setResolvedOllamaHost(OLLAMA_LOCALHOST);
+    const { capture, calls } = makeOllamaCapture([
+      { match: /\/api\/tags/, output: JSON.stringify({ models: [] }) },
+      {
+        match: /ollama list/,
+        output:
+          "NAME           ID            SIZE    MODIFIED\nllama3.2:3b    abc123        2.0 GB  2 days ago\n",
+      },
+    ]);
+    expect(getOllamaModelOptions(capture)).toEqual(["llama3.2:3b"]);
+    expect(calls.some((argv) => argv.includes("list"))).toBe(true);
   });
 
-  it("prefers Ollama /api/tags over parsing the CLI list output", () => {
-    let call = 0;
-    const mockCapture = () => {
-      call += 1;
-      if (call === 1) {
-        return JSON.stringify({ models: [{ name: "qwen3.5:9b" }] });
-      }
-      return "";
-    };
-    expect(getOllamaModelOptions(mockCapture)).toEqual(["qwen3.5:9b"]);
+  it("returns parsed tags without calling the loopback CLI", () => {
+    setResolvedOllamaHost(OLLAMA_HOST_DOCKER_INTERNAL);
+    const { capture, calls } = makeOllamaCapture([
+      {
+        match: /\/api\/tags/,
+        output: JSON.stringify({ models: [{ name: "qwen3.5:9b" }, { name: "gemma2:9b" }] }),
+      },
+    ]);
+    expect(getOllamaModelOptions(capture)).toEqual(["qwen3.5:9b", "gemma2:9b"]);
+    expect(calls).toHaveLength(1);
   });
 
-  it("returns no installed ollama models when list output is empty", () => {
-    let call = 0;
-    const mockCapture = () => {
-      call += 1;
-      return call === 1 ? JSON.stringify({ models: [] }) : "";
-    };
-    expect(getOllamaModelOptions(mockCapture)).toEqual([]);
+  it("retries an invalid Windows-host inventory before returning installed models (#10259)", () => {
+    setResolvedOllamaHost(OLLAMA_HOST_DOCKER_INTERNAL);
+    const outputs = [
+      "",
+      "<html>proxy response</html>",
+      JSON.stringify({ models: [{ name: "qwen3.5:9b" }] }),
+    ];
+    const capture = vi.fn(() => outputs.shift() ?? "");
+    const sleeps: number[] = [];
+    const models = getOllamaModelOptions(capture, (milliseconds) => sleeps.push(milliseconds));
+    expect(models).toEqual(["qwen3.5:9b"]);
+    expect(capture).toHaveBeenCalledTimes(3);
+    expect(sleeps).toEqual([500, 1_000]);
+  });
+
+  it("rejects an invalid Windows-host inventory after bounded retries (#10259)", () => {
+    setResolvedOllamaHost(OLLAMA_HOST_DOCKER_INTERNAL);
+    const capture = vi.fn(() => "");
+    const sleeps: number[] = [];
+    expect(() =>
+      getOllamaModelOptions(capture, (milliseconds) => sleeps.push(milliseconds)),
+    ).toThrow(/Could not read Ollama models from host\.docker\.internal:11434 after 3 attempts/);
+    expect(capture).toHaveBeenCalledTimes(3);
+    expect(sleeps).toEqual([500, 1_000]);
   });
 
   it("prefers the default ollama model when present", () => {
